feat(bottle): add web view APIs for my
